# docs/swagger.yaml
definitions:
...
  github_com_PritOriginal_problem-map-server_internal_models.Check:
    properties:
      check_id:
        type: integer
      comment:
        type: string
      created_at:
        type: string
      mark_id:
        type: integer
      mark_status_history_id:
        type: integer
      mark_status_id:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.MarkStatusType'
      photos:
        items:
          type: string
        type: array
      result:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.MarkStatusType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    type: integer
    x-enum-varnames:
    - UnconfirmedStatus
    - ConfirmedStatus
    - UnderReviewStatus
    - RediscoveredStatus
    - ClosedStatus
    - RefutedStatus
    - InProgressStatus
    - DuplicateStatus
...
  github_com_PritOriginal_problem-map-server_internal_models.Notification:
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: integer
      mark_id:
        type: integer
      read_at:
        format: date-time
        type: string
      task_id:
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.NotificationType'
      user_id:
        type: integer
    type: object
  github_com_PritOriginal_problem-map-server_internal_models.NotificationType:
    enum:
    - mark.status_changed
    - task.assigned
    - check.added
    - mark.assigned
    - mark.sla_breached
    - comment_added
    - mark.hidden
    - mark.merged
    - webhook.disabled
    - badge_earned
    type: string
    x-enum-varnames:
    - NotificationMarkStatusChanged
    - NotificationTaskAssigned
    - NotificationCheckAdded
    - NotificationMarkAssigned
    - NotificationMarkSLABreached
    - NotificationCommentAdded
    - NotificationMarkHidden
    - NotificationMarkMerged
    - NotificationWebhookDisabled
    - NotificationBadgeEarned
...
  github_com_PritOriginal_problem-map-server_internal_models.PointJSON:
    properties:
      coordinates:
        example:
        - 0
        - 0
        items:
          type: number
        type: array
      type:
        type: string
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.Role:
    enum:
    - user
    - moderator
    - admin
    - service
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleModerator
    - RoleAdmin
    - RoleService
...
  github_com_PritOriginal_problem-map-server_internal_models.Task:
    properties:
      created_at:
        type: string
      due_at:
        description: |-
          DueAt is the deadline set by the tasker; tasks still issued after it
          are moved to OverdueStatus by Tasker.ExpireOverdue.
        format: date-time
        type: string
      mark_id:
        type: integer
      name:
        type: string
      status_id:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.TaskStatusType'
      task_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.TaskStatusType:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - UnfulfilledStatus
    - CompletedStatus
    - OverdueStatus
...
  github_com_PritOriginal_problem-map-server_internal_models.User:
    properties:
      created_at:
        description: CreatedAt is the sign-up time (users.created_at).
        type: string
      home_point:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.PointJSON'
      login:
        type: string
      rating:
        type: integer
      role:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.Role'
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.UserSyncTotals:
    properties:
      checks:
        type: integer
      notifications:
        type: integer
      tasks:
        type: integer
    type: object
...
  github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo:
    properties:
      message:
        type: string
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-any:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload: {}
      success:
        type: boolean
    type: object
...
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_sync_GetUserSyncResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_sync.GetUserSyncResponse'
      success:
        type: boolean
    type: object
...
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_users_GetMeResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_users.GetMeResponse'
      success:
        type: boolean
    type: object
...
  internal_handler_sync.GetUserSyncResponse:
    properties:
      checks:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.Check'
        type: array
      notifications:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.Notification'
        type: array
      server_time:
        type: string
      tasks:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.Task'
        type: array
      totals:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.UserSyncTotals'
    type: object
...
  internal_handler_users.GetMeResponse:
    properties:
      user:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.User'
    type: object
...
info:
  contact: {}
  description: This is the API documentation for the "Problem Map" project.
  title: Problem Map API
  version: "1.0"
paths:
...
  /users/me:
    get:
      description: get full profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_users_GetMeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
...
  /users/me/sync:
    get:
      description: 'one request for a client coming back online: the caller''s tasks
        updated after `since`, unread notifications received after it and checks submitted
        after it. Each collection is paged independently by limit/offset; `totals`
        carry the full counts. Store `server_time` and pass it as the next `since`;
        with several server instances subtract a safety margin (`server_time - 1s`)
        to cover clock skew between them. `since` in the future is rejected with 400'
      parameters:
      - description: instant of the previous sync (RFC3339 or YYYY-MM-DD)
        in: query
        name: since
        required: true
        type: string
      - default: 100
        description: page size of each collection, 1..500
        in: query
        name: limit
        type: integer
      - default: 0
        description: page offset of each collection
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_sync_GetUserSyncResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - BearerAuth: []
      summary: Personal changes since an instant
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Open-data API key "pm_live_…" (POST /api-keys); also accepted as
      "Authorization: ApiKey {key}". Optional on public GET routes: with a key the
      per-key quota applies and X-RateLimit-* headers are returned; ignored when a
      valid Bearer JWT is sent as well (the JWT identity wins)'
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'JWT access token: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...